format changes, converted to float and organized imports

+++ test-engineering/contract-tests/client/tests/models.py
@@ -22,7 +22,7 @@ class Request(BaseModel):
     path: str
     headers: List[Header] = []
     # Delay is optional, providing time for data refresh
-    delay: Optional[int] = 0
+    delay: Optional[float] = 0
 
 
 class Suggestion(BaseModel, extra=Extra.allow):
